+++ packages/base/Helpbox/src/Helpbox/Helpbox.tsx
@@ -1,15 +1,12 @@
 import type { ReactNode, HTMLAttributes } from 'react'
 import React, { forwardRef } from 'react'
-import type { Theme } from '@material-ui/core/styles'
-import { makeStyles } from '@material-ui/core/styles'
-import cx from 'classnames'
 import type { BaseProps } from '@toptal/picasso-shared'
 import { Container } from '@toptal/picasso-container'
 import { CloseMinor16 } from '@toptal/picasso-icons'
 import { ButtonCircular } from '@toptal/picasso-button'
 import type { VariantType as ContainerVariantType } from '@toptal/picasso-container'
+import { twMerge } from '@toptal/picasso-tailwind-merge'
 
-import styles from './styles'
 import HelpboxContext from './HelpboxContext'
 
 export interface Props extends BaseProps, HTMLAttributes<HTMLDivElement> {
@@ -21,24 +18,19 @@ export interface Props extends BaseProps, HTMLAttributes<HTMLDivElement> {
   onClose?: (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => void
 }
 
-const useStyles = makeStyles<Theme>(styles, {
-  name: 'PicassoHelpbox',
-})
-
 // eslint-disable-next-line react/display-name
 export const Helpbox = forwardRef<HTMLDivElement, Props>(function Helpbox(
   props,
   ref
 ) {
   const { className, style, children, variant, onClose, ...rest } = props
-  const classes = useStyles()
 
   return (
     <Container
       {...rest}
       ref={ref}
       rounded
-      className={cx(classes.root, className)}
+      className={twMerge('relative', className)}
       style={style}
       bordered={!variant || variant === 'white'}
       variant={variant}