+++ packages/base/Page/src/PageTopBarMenu/PageTopBarMenu.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable complexity */
 import type { ReactNode, HTMLAttributes } from 'react'
-import React, { forwardRef } from 'react'
+import React, { useContext, forwardRef } from 'react'
 import cx from 'classnames'
 import type { Theme } from '@material-ui/core/styles'
 import { makeStyles } from '@material-ui/core/styles'
@@ -12,6 +12,7 @@ import { Typography } from '@toptal/picasso-typography'
 import { DropdownCompound as Dropdown } from '@toptal/picasso-dropdown'
 
 import styles from './styles'
+import { PageTopBarContext } from '../PageTopBar'
 
 const useStyles = makeStyles<Theme>(styles, {
   name: 'PicassoTopBarMenu',
@@ -42,13 +43,16 @@ export const PageTopBarMenu = forwardRef<HTMLDivElement, Props>(
     } = props
     const classes = useStyles()
 
+    const { variant } = useContext(PageTopBarContext)
+    const invert = variant === 'light'
+
     const isCompactLayout = useBreakpoint(['xs', 'sm', 'md'])
 
     const metaContent =
       typeof meta === 'string' ? (
         <Typography
           className={classes.truncateText}
-          invert={!isCompactLayout}
+          invert={!isCompactLayout && !invert}
           size='xsmall'
           data-private={dataPrivate}
         >
@@ -89,7 +93,7 @@ export const PageTopBarMenu = forwardRef<HTMLDivElement, Props>(
       />
     ) : (
       <UserBadge
-        invert
+        invert={!invert}
         center
         size='xxsmall'
         classes={{
@@ -122,7 +126,9 @@ export const PageTopBarMenu = forwardRef<HTMLDivElement, Props>(
         }}
       >
         {trigger}
-        <Dropdown.Arrow className={classes.arrow} />
+        <Dropdown.Arrow
+          className={cx(classes.arrow, { [classes.invert]: invert })}
+        />
       </Dropdown>
     )
   }